apps/controller/requirements.txt
kubernetes
PyYAML>=6.0
Jinja2>=3.1
